+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-04-29 19:21:06
-// Commit: feat(storage): integrate custom telemetry logger in storage
+// Updated at: 2026-05-02 16:22:35
+// Commit: feat(storage): add bi-directional event bus handler for storage
 
 class storageComponentState {
-  final String commitRef = '773';
-  final String lastMessage = 'feat(storage): integrate custom telemetry logger in storage';
-  final DateTime timestamp = DateTime.parse('2026-04-29T19:21:06+05:00');
+  final String commitRef = '791';
+  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
+  final DateTime timestamp = DateTime.parse('2026-05-02T16:22:35+05:00');
 }